requirements.txt
lxml==5.3.0
numpy==2.2.2
pandas==2.2.3
pillow==11.1.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
pytz==2024.2
six==1.17.0
typing_extensions==4.12.2
tzdata==2025.1
